requirements.txt
python-dotenv>=1.0.0
huggingface_hub>=0.28.0
